pyproject.toml
[project]
name = "playwright-tests"
version = "0.1.0"
description = ""
authors = [
    {name = "yulichka.metel",email = "yulichka.metel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright (>=1.52.0,<2.0.0)",
    "pytest-playwright (>=0.7.0,<0.8.0)",
    "pillow (>=11.2.1,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
